pyproject.toml
[project]
name = "story-crafter-agent"
version = "0.1.0"
description = "ADK Agent for Creating Personalized Illustrated Stories from Classic Literature"
requires-python = ">=3.10,<3.14"
dependencies = [
    "google-adk>=1.18.0",
    "google-generativeai>=0.8.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
    "langchain-community>=0.4.1",
    "wikipedia>=1.4.0",
    "mcp>=1.21.1",
    "httpx>=0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["story_crafter_agent"]

